+++ debug/plugins/prism-es6/components/prism-elixir.js
@@ -724,6 +724,11 @@ Prism.languages.elixir = {
 
 
 
+
+
+
+
+
       // See interpolation below
     } }, { pattern: /("""|''')[\s\S]*?\1/, greedy: true, inside: {