+++ debug/plugins/sweetalert2/src/SweetAlert.js
@@ -998,6 +998,11 @@ location.host.match(/\.(ru|su|by|xn--p1ai)$/))
 
 
 
+
+
+
+
+
           // ignore
         });}, 2500);}, 500);}} // Assign instance methods from src/instanceMethods/*.js to prototype
 SweetAlert.prototype.disableButtons = instanceMethods.disableButtons;SweetAlert.prototype.enableButtons = instanceMethods.enableButtons;SweetAlert.prototype.getInput = instanceMethods.getInput;SweetAlert.prototype.disableInput = instanceMethods.disableInput;SweetAlert.prototype.enableInput = instanceMethods.enableInput;SweetAlert.prototype.hideLoading = instanceMethods.hideLoading;SweetAlert.prototype.disableLoading = instanceMethods.disableLoading;SweetAlert.prototype.showValidationMessage = instanceMethods.showValidationMessage;SweetAlert.prototype.resetValidationMessage = instanceMethods.resetValidationMessage;SweetAlert.prototype.close = instanceMethods.close;SweetAlert.prototype.closePopup = instanceMethods.closePopup;SweetAlert.prototype.closeModal = instanceMethods.closeModal;SweetAlert.prototype.closeToast = instanceMethods.closeToast;SweetAlert.prototype.rejectPromise = instanceMethods.rejectPromise;SweetAlert.prototype.update = instanceMethods.update;SweetAlert.prototype._destroy = instanceMethods._destroy; // Assign static methods from src/staticMethods/*.js to constructor